requirements.txt
customtkinter
Pillow
pycryptodome
opencv-python
numpy
mss
screeninfo